1개의 게시물 접근 상세 정보 조회 + 해당 기능들을 수행하기 위한 모든 dao 설정

+++ src/main/java/com/team2/docgram/dao/UserDaoImpl.java
@@ -32,7 +32,7 @@ public class UserDaoImpl implements UserDao {
 	 * @since 2022-05-19
 	 */
 	@Override
-	public UserDto readOne(Integer num) {
+	public UserDto readOne(Long num) {
 		return sqlSession.selectOne(mapper+"readOne", num);
 	}
